Remove crosssearch module
