+++ launchii/appsearch.py
@@ -2,14 +2,28 @@ import os
 from collections import OrderedDict
 import getpass
 import json
-from launchii.crosssearch import BaseSearch
 
 
-class StartMenuSearch(BaseSearch):
-    
+class StartMenuSearch:
     @staticmethod
     def supported_environment(platform: str) -> bool:
         return platform == "Windows"
+
+    def __init__(self):
+        with open("index.json", "r") as f:  # load index
+            self.index = json.load(f)
+
+    def search(self, search_term) -> dict:
+        results = {}
+        for i in self.index:  # iterate over index
+            if search_term in self.index[i].lower():  # if search term is in index
+                # append i to results as key and index[i] as value
+                results[i] = self.index[i]
+        return results
+
+    def get_path(self, term) -> str:
+        path = self.index[term]
+        return path
 
     def createindex(
         self,
