+++ launchii/macappsearch.py
@@ -3,7 +3,6 @@ from collections import OrderedDict
 import getpass
 import json
 from glob import glob
-from launchii.crosssearch import BaseSearch
 
 path = [
     "/Applications",
@@ -13,11 +12,26 @@ path = [
 rawFileList = {}
 
 
-class OSXApplicationSearch(BaseSearch):
-    
+class OSXApplicationSearch:
     @staticmethod
     def supported_environment(platform: str) -> bool:
         return platform == "Darwin"
+
+    def __init__(self):
+        with open("index.json", "r") as f:  # load index
+            self.index = json.load(f)
+
+    def search(self, search_term) -> dict:
+        results = {}
+        for i in self.index:  # iterate over index
+            if search_term in self.index[i].lower():  # if search term is in index
+                # append i to results as key and index[i] as value
+                results[i] = self.index[i]
+        return results
+
+    def get_path(self, term) -> str:
+        path = self.index[term]
+        return path
 
     def createIndex(self, path=path) -> OrderedDict:
         # For every filepath in path, find all the folders (no subdirectories) ending in .app and add them to the dict rawFileList
